chore: build process revamped completely moved rollup out
