+++ .changeset/sour-otters-trade.md
@@ -1,0 +1,5 @@
+---
+"rehype-code-group": minor
+---
+
+chore: build process revamped for unminifed dist
