+++ package.json
@@ -2,21 +2,20 @@
   "name": "rehype-code-group",
   "version": "0.1.3",
   "description": "A Rehype plugin for grouping code blocks with tabs, perfect for documentation and tutorials. Works with all Code Syntax Highlighters.",
-  "main": "dist/index.cjs",
-  "module": "dist/index.mjs",
-  "types": "dist/index.d.ts",
-  "type": "module",
-  "files": ["dist", "src"],
+  "main": "dist/index.js",
+  "module": "dist/index.js",
   "exports": {
     ".": {
-      "import": "./dist/index.mjs",
-      "require": "./dist/index.cjs",
+      "import": "./dist/index.js",
       "types": "./dist/index.d.ts"
     }
   },
+  "types": "dist/index.d.ts",
+  "type": "module",
+  "files": ["dist"],
   "scripts": {
     "run:example": "pnpm build && node example/index.js",
-    "build": "pnpm clean && rollup -c --bundleConfigAsCjs",
+    "build": "pnpm clean && tsc",
     "clean": "rimraf dist",
     "format": "biome format --write",
     "lint": "biome check --fix",
@@ -37,17 +36,12 @@
     "@biomejs/biome": "1.9.4",
     "@changesets/changelog-github": "^0.5.0",
     "@changesets/cli": "^2.27.9",
-    "@rollup/plugin-commonjs": "^28.0.1",
-    "@rollup/plugin-node-resolve": "^15.3.0",
-    "@rollup/plugin-terser": "^0.4.4",
-    "@rollup/plugin-typescript": "^12.1.1",
     "@types/hast": "^3.0.4",
     "@types/jest": "^29.5.14",
     "@types/node": "^22.9.0",
     "jest": "^29.7.0",
     "lint-staged": "^15.2.10",
     "rimraf": "^6.0.1",
-    "rollup": "^4.25.0",
     "simple-git-hooks": "^2.11.1",
     "ts-jest": "^29.2.5",
     "ts-node": "^10.9.2",
