--- rollup.config.js
@@ -1,29 +1,0 @@
-import commonjs from "@rollup/plugin-commonjs";
-import resolve from "@rollup/plugin-node-resolve";
-import terser from "@rollup/plugin-terser";
-import typescript from "@rollup/plugin-typescript";
-
-const external = [...Object.keys(require("./package.json").dependencies || {})];
-
-export default {
-  input: "src/index.ts",
-  output: [
-    {
-      file: "dist/index.cjs",
-      format: "cjs",
-    },
-    {
-      file: "dist/index.mjs",
-      format: "esm",
-    },
-  ],
-  plugins: [
-    resolve(),
-    commonjs(),
-    typescript({
-      tsconfig: "./tsconfig.json",
-    }),
-    terser(),
-  ],
-  external,
-};
