+++ tsconfig.json
@@ -1,34 +1,27 @@
 {
   "compilerOptions": {
     // Enable latest features
-    "lib": ["ESNext"],
+    "lib": ["ESNext", "DOM"],
     "target": "ESNext",
-    "module": "ESNext",
+    "module": "NodeNext",
     "moduleDetection": "force",
-
-    // Bundler mode
-    "moduleResolution": "bundler",
     "verbatimModuleSyntax": true,
-    "noEmit": true,
-
     // Best practices
     "strict": true,
     "skipLibCheck": true,
     "noFallthroughCasesInSwitch": true,
     "esModuleInterop": true,
-
     // Some stricter flags (disabled by default)
-    "noUnusedLocals": false,
-    "noUnusedParameters": false,
-    "noPropertyAccessFromIndexSignature": false,
+    "noUnusedLocals": true,
+    "noUnusedParameters": true,
+    "noPropertyAccessFromIndexSignature": true,
 
     "rootDir": "./src",
     "outDir": "./dist",
     "forceConsistentCasingInFileNames": true,
 
     "declaration": true,
-    "declarationMap": true,
-    "declarationDir": "./dist"
+    "declarationMap": true
   },
   "include": ["src"],
   "exclude": ["node_modules", "dist"]
